feat: use assets and add getNewsFromAssets way
