+++ app/src/main/java/com/weiran/mynowinandroid/data/source/LocalStorageImpl.kt
@@ -1,8 +1,6 @@
 package com.weiran.mynowinandroid.data.source
 
 import android.content.Context
-import android.content.res.AssetManager
-import android.util.Log
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 import com.weiran.mynowinandroid.R
@@ -14,9 +12,6 @@ import com.weiran.mynowinandroid.utils.FileUtil
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import java.io.BufferedReader
-import java.io.IOException
-import java.io.InputStreamReader
 import javax.inject.Inject
 
 class LocalStorageImpl @Inject constructor(
@@ -26,23 +21,12 @@ class LocalStorageImpl @Inject constructor(
     private val gson = Gson()
 
     //从文件中读取内容转成json字符串
-    private fun getNewsFromAssets(fileName: String): String {
-        val stringBuilder = StringBuilder()
-        return try {
-            val assetManager: AssetManager = context.assets
-            val isr = InputStreamReader(assetManager.open(fileName))
-            val bf = BufferedReader(isr)
-            var line: String?
-            while (bf.readLine().also { line = it } != null) {
-                stringBuilder.append(line)
-            }
-            bf.close()
-            isr.close()
-            stringBuilder.toString()
-        } catch (e: IOException) {
-            Log.e("getNewsFromAssets Error", e.printStackTrace().toString())
-            ""
-        }
+    override fun getNewsFromAssets(): List<News> {
+        return gson.fromJson(
+            FileUtil.getNewsFromAssets(
+                context, ASSETS_NAME
+            ), object : TypeToken<List<News>>() {}.type
+        )
     }
 
 
@@ -80,6 +64,10 @@ class LocalStorageImpl @Inject constructor(
                 appDatabase.topicDao().insert(topicEntity)
             }
         }
+    }
+
+    companion object {
+        private const val ASSETS_NAME = "news.json"
     }
 
 }
