+++ app/src/main/java/com/weiran/mynowinandroid/utils/FileUtil.kt
@@ -1,8 +1,12 @@
 package com.weiran.mynowinandroid.utils
 
 import android.content.Context
+import android.content.res.AssetManager
+import android.util.Log
 import androidx.annotation.RawRes
+import java.io.BufferedReader
 import java.io.IOException
+import java.io.InputStreamReader
 
 object FileUtil {
 
@@ -17,5 +21,26 @@ object FileUtil {
             str
         }
     }
+
+    fun getNewsFromAssets(context: Context, assetsName: String): String {
+        val stringBuilder = StringBuilder()
+        return try {
+            val assetManager: AssetManager = context.assets
+            val inputStreamReader = InputStreamReader(assetManager.open(assetsName))
+            val bufferedReader = BufferedReader(inputStreamReader)
+            var line: String?
+            while (bufferedReader.readLine().also { line = it } != null) {
+                stringBuilder.append(line)
+            }
+            bufferedReader.close()
+            inputStreamReader.close()
+            stringBuilder.toString()
+        } catch (e: IOException) {
+            Log.e(TAG, e.printStackTrace().toString())
+            ""
+        }
+    }
+
+    private const val TAG = "FileUtil"
 
 }
